Child Learning System Using Assembly Language
This is a minimalistic children's learning application implemented in Assembly language. The program aims to help young learners, such as children or beginners, in their early educational journey by providing an interactive platform to explore alphabets and digits. It prompts users to select between two options: "STUDY" and "Exit," using a text-based menu.